The beginning of a CGI example.
Brigitte wants an example, to clarify, and of course she's right.
But could I be putting too many details in this pseudo-implementation?
I want to be as flexible as possible for the future examples based on it,
but I will really need to trim the code down to the minimum once I know
which parts I'll need.